refactored app structure
